Using BigDecimal for stop loss and profit limit

+++ src/main/java/com/robindrew/trading/position/order/IPositionOrder.java
@@ -16,8 +16,8 @@ public interface IPositionOrder {
 
 	BigDecimal getTradeSize();
 
-	int getProfitLimitDistance();
+	BigDecimal getProfitLimitDistance();
 
-	int getStopLossDistance();
+	BigDecimal getStopLossDistance();
 
 }
